Added code to allow usage of the split screen menu button and the start/stop online button. With light testing, allows a player to start an online session with a splitscreen partner active. Also allows a splitscreen partner to join a player that is already connected to an online lobby.
Find game is not allowed for an active split screen session as it resulted in both characters losing almost all of their inventory on joining. Disconnecting from an online session with a  splitscreen partner active is not allowed due to a spawning bug that spawns the splitscreen player at the last place the main player entered that zone, and not by the side of the main player.

Ideal usage is host and splitscreen player open game to friends, remote player joins alone, then remote splitscreen player joins. Ideal is also that remote splitscreen player disconnects before the main remote player does.